requirements.txt
Flask==0.9
SQLAlchemy==0.7.8
Flask-Login==0.1
Flask-WTF==0.6
Flask-Gravatar
Flask-OAuth==0.12
Flask-SQLAlchemy
Flask-Mail
dateutils
alembic
psycopg2
requests
itsdangerous
markdown
raven
blinker
flask-cache
europeana-search
flask-heroku
pygeoip
flask-babel
